Light refactors to shared tasks, fixes to align promotion to dataset refactors
