+++ dags/veda_data_pipeline/veda_dataset_pipeline.py
@@ -1,12 +1,11 @@
 import pendulum
 from airflow import DAG
-from copy import deepcopy
 from airflow.models.param import Param
 from airflow.decorators import task
 from veda_data_pipeline.groups.discover_group import discover_from_s3_task, get_files_task
 from airflow.operators.dummy_operator import DummyOperator as EmptyOperator
 from veda_data_pipeline.groups.collection_group import collection_task_group
-from veda_data_pipeline.groups.processing_tasks import submit_to_stac_ingestor_task, build_stac_task
+from veda_data_pipeline.groups.processing_tasks import submit_to_stac_ingestor_task, build_stac_task, extract_discovery_items_from_payload, remove_thumbnail_asset
 
 template_dag_run_conf = {
     "collection": "<collection-id>",
@@ -50,29 +49,9 @@ with DAG("veda_dataset_pipeline", params=template_dag_run_conf, **dag_args) as d
     start = EmptyOperator(task_id="start")
     end = EmptyOperator(task_id="end")
 
-
-    @task()
-    def remove_thumbnail_asset(ti):
-        payload = deepcopy(ti.dag_run.conf)
-        payloads = list()
-        assets = payload.get("assets", {})
-        if assets.get("thumbnail"):
-            assets.pop("thumbnail")
-        # if thumbnail was only asset, delete assets
-        if not assets:
-            payload.pop("assets")
-        for item in payload.get("discovery_items"):
-            payloads.append({
-                **payload,
-                **item
-            }
-            )
-
-        return payloads
-
-
     mutated_payloads = start >> collection_task_group() >> remove_thumbnail_asset()
-    discover = discover_from_s3_task.expand(payload=mutated_payloads)
+    discovery_items = extract_discovery_items_from_payload(mutated_payloads)
+    discover = discover_from_s3_task.partial(payload=mutated_payloads).expand(event=discovery_items)
     get_files = get_files_task(payload=discover)
     build_stac = build_stac_task.expand(payload=get_files)
     submit_stac = submit_to_stac_ingestor_task.expand(built_stac=build_stac) >> end
